requirements.txt
Django==2.2.17
djangorestframework==3.12.2
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.4
sqlparse==0.4.1
